+++ tests/Unit/Schema/Directives/DeprecatedDirectiveTest.php
@@ -5,6 +5,7 @@ namespace Tests\Unit\Schema\Directives;
 use Tests\TestCase;
 use Illuminate\Support\Arr;
 use Tests\Utils\Resolvers\Foo;
+use GraphQL\Type\Definition\Directive;
 
 class DeprecatedDirectiveTest extends TestCase
 {
@@ -21,6 +22,8 @@ class DeprecatedDirectiveTest extends TestCase
                 @deprecated(reason: \"{$reason}\") 
                 @field(resolver: \"{$resolver}\")
             bar: String
+                @field(resolver: \"{$resolver}\")
+            withDefaultReason: String
                 @field(resolver: \"{$resolver}\")
         }
         ";
@@ -60,8 +63,17 @@ class DeprecatedDirectiveTest extends TestCase
                 return $field['isDeprecated'];
             }
         );
-        $this->assertCount(1, $deprecatedFields);
-        $this->assertSame($reason, $deprecatedFields[0]['deprecationReason']);
+        $this->assertCount(2, $deprecatedFields);
+        $this->assertSame(
+            $reason,
+            $deprecatedFields[0]['deprecationReason'],
+            'Should show user-defined deprecation reason.'
+        );
+        $this->assertSame(
+            Directive::DEFAULT_DEPRECATION_REASON,
+            $deprecatedFields[1]['deprecationReason'],
+            'Should fallback to the default deprecation reason'
+        );
 
         $this->query('
         {