+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/nuwave/lighthouse/compare/v3.6.0...master)
 
+### Fixed
+
+- Use the spec-compliant default deprecation reason for `@deprecate` directive
+
 ## [3.6.0](https://github.com/nuwave/lighthouse/compare/v3.5.3...v3.6.0)
 
 ### Added